Add tls support as a cryption layer
    Traditionally, goproxy used cryptconn as its cryption layer. It's a
self-made protocol. I made it cause I concern about TLS. It's so widely used
that something act not like https will be discovered. But it was proofed that
I was wrong. Using TLS will reduce the possibility be founded. So I added it as
an option. User can choice TLS or cryptconn by his own situation.

  + tls listener & dialer
  * split client.go from server.go
  * move client & server's config to them own file