fix memory leak issue
    Normally Tunnel don't care frame's delivery, that's Okey. However, When a
connection trying to close, and the tunnel is broken. That's connection will
hold on the close-waiting state until the tcp connection under tunnel are
finally timeout. That normally will take 924s. In this period, all
connection will not be released, that cause memory leak.
    I found this when my router re-dial and got another outside IP address.
Goproxy know nothing about this but connection are broken. As a result,
I found everything blocked, lots of connections keep in CLOSEWAIT, and
goproxy's memory abnormal.